refactor: Set Appium as peer dependency
