+++ ci-jobs/ci.yml
@@ -1,7 +1,7 @@
 jobs:
   - job: EspressoCI
     pool:
-      vmImage: macOS-10.15
+      vmImage: macOS-12
     variables:
       CI: true
       TERM: dumb
@@ -10,7 +10,7 @@ jobs:
     steps:
     - task: NodeTool@0
       inputs:
-        versionSpec: 14.x
+        versionSpec: 16.x
     - task: JavaToolInstaller@0
       inputs:
         versionSpec: '11'
