+++ ci-jobs/templates/android-e2e-template.yml
@@ -1,10 +1,10 @@
 # https://docs.microsoft.com/azure/devops/pipelines/languages/android
 parameters:
   name: 'android_e2e_tests'
-  vmImage: 'macOS-10.15'
+  vmImage: 'macOS-12'
   MOCHA_FILE: 'mocha-test-results.xml'
   CHROMEDRIVER_VERSION: 2.44
-  NODE_VERSION: 14.x
+  NODE_VERSION: 16.x
   ANDROID_SDK_VERSION: 28
 
 jobs:
