+++ lib/commands/execute.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 const extensions = {};
