+++ lib/commands/general.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 import validate from 'validate.js';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 import { qualifyActivityName } from '../utils';
 
 const commands = {}, helpers = {}, extensions = {};
