+++ lib/driver.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 import path from 'path';
-import { BaseDriver, errors, isErrorType, DeviceSettings} from '@appium/base-driver';
+import { BaseDriver, errors, isErrorType, DeviceSettings} from 'appium/driver';
 import { EspressoRunner, TEST_APK_PKG } from './espresso-runner';
-import { fs, tempDir, zip } from '@appium/support';
+import { fs, tempDir, zip } from 'appium/support';
 import commands from './commands';
 import { DEFAULT_ADB_PORT } from 'appium-adb';
 import { androidHelpers, androidCommands, SETTINGS_HELPER_PKG_ID } from 'appium-android-driver';
