+++ lib/espresso-runner.js
@@ -1,8 +1,8 @@
-import { JWProxy, errors } from '@appium/base-driver';
+import { JWProxy, errors } from 'appium/driver';
 import { waitForCondition } from 'asyncbox';
 import { ServerBuilder, buildServerSigningConfig } from './server-builder';
 import path from 'path';
-import { fs, util, mkdirp, timing } from '@appium/support';
+import { fs, util, mkdirp, timing } from 'appium/support';
 import { version } from '../../package.json'; // eslint-disable-line import/no-unresolved
 import B from 'bluebird';
 import _ from 'lodash';
