+++ lib/server-builder.js
@@ -1,5 +1,5 @@
 import { SubProcess } from 'teen_process';
-import { fs, system } from '@appium/support';
+import { fs, system } from 'appium/support';
 import _ from 'lodash';
 import path from 'path';
 import { EOL } from 'os';
