+++ lib/utils.js
@@ -1,4 +1,4 @@
-import { fs, mkdirp } from '@appium/support';
+import { fs, mkdirp } from 'appium/support';
 import _ from 'lodash';
 import path from 'path';
