+++ package.json
@@ -49,8 +49,6 @@
     "!.DS_Store"
   ],
   "dependencies": {
-    "@appium/base-driver": "^8.5.2",
-    "@appium/support": "^2.58.0",
     "@babel/runtime": "^7.4.3",
     "appium-adb": "^9.1.0",
     "appium-android-driver": "^5.0.0",
@@ -83,11 +81,16 @@
     "precommit-msg",
     "precommit-test"
   ],
+  "peerDependencies": {
+    "appium": "^2.0.0-beta.40"
+  },
   "devDependencies": {
     "@appium/gulp-plugins": "^7.0.0",
     "@appium/eslint-config-appium": "^6.0.0",
     "@semantic-release/git": "^10.0.1",
     "@xmldom/xmldom": "^0.x",
+    "appium": "^2.0.0-beta.40",
+    "appium-espresso-driver": "file:.",
     "android-apidemos": "^4.0.0",
     "appium-chromedriver": "^5.0.0",
     "async-lock": "^1.0.0",
